Updated these scripts to reflect the changes in CatHealth script from changing float health to int

+++ Assets/Scripts/Cloud/CloudAttack.cs
@@ -6,7 +6,7 @@ using UnityEngine.Events;
 public class CloudAttack : MonoBehaviour
 {
     [SerializeField]
-    private float _attackDamage;
+    private int _attackDamage;
 
     private UnityAction OnCloudAttack;
